Add basic "zero state" when no search results
On the /search page, if there are no search results, show a grey box
that says "No annotations matched your search."

If on the /groups/<pubid>/search page and there is no search query
(other than the group) then the message changes to 'The group "{name}
have not made any annotations yet." When there is a search query other
than just the group the original  "No annotations matched your search"
message is still used.

If on the /users/<username>/search page and there is no search query
(other than the user) then the message changes to "{name} hasn't made
any annotations yet". If this is the search page of the authorized user
then "You haven't made any annotations yet" is used. When there is a search
query other than just the user the original  "No annotations matched your
search" message is still used.